small refactoring + making it clearer
