+++ caribou/client/src/pages/Map.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import styled, { keyframes } from "styled-components";
+import styled from "styled-components";
 import {
   GoogleMap,
   InfoWindow,
@@ -7,7 +7,8 @@ import {
   Marker,
 } from "@react-google-maps/api";
 import jb from "../assets/jb.png";
-import { Button, Title } from "../StyledComponents";
+import { Button, Title } from "../components/index.js";
+import MapModal from "../MapModal";
 
 const containerStyle = {
   width: "350px",
@@ -28,17 +29,17 @@ const Map = () => {
   const [excitementLevel, setExcitementLevel] = useState("");
   const [humanPosition, setHumanPosition] = useState(null);
   const [center, setCenter] = useState({
-    //default center is montreal
+    //default center of map is middle of montreal
     lat: 45.5017,
     lng: -73.5673,
   });
+  useEffect(() => {
+    //gets info on all those pesky humans
+    fetch("/humans")
+      .then((res) => res.json())
+      .then((data) => setHumans(data.data));
+  }, []);
 
-  const handleTrash = (event) => {
-    setTrashingLevel(event.target.value.replace(/[^0-9]/g, ""));
-  };
-  const handleExcitement = (event) => {
-    setExcitementLevel(event.target.value.replace(/[^0-9]/g, ""));
-  };
   const handleDeleteHuman = () => {
     fetch("/deletehuman", {
       method: "delete",
@@ -60,7 +61,7 @@ const Map = () => {
       });
   };
   const handleAddHuman = () => {
-    const newOne = {
+    const newHuman = {
       position: humanPosition,
       date: new Date(),
       trashingLevel: +trashingLevel,
@@ -71,7 +72,7 @@ const Map = () => {
       headers: {
         "Content-Type": "application/json",
       },
-      body: JSON.stringify(newOne),
+      body: JSON.stringify(newHuman),
     })
       .then((res) => res.json())
       .then((data) => {
@@ -79,7 +80,7 @@ const Map = () => {
           setTrashingLevel("");
           setExcitementLevel("");
           setHumanModal(false);
-          setHumans([...humans, newOne]);
+          setHumans([...humans, newHuman]);
         } else {
           console.log("ERROR");
         }
@@ -89,38 +90,17 @@ const Map = () => {
     setHumanPosition({ lat: event.latLng.lat(), lng: event.latLng.lng() });
     setHumanModal(true);
   };
-  useEffect(() => {
-    //gets info on all those pesky humans
-    fetch("/humans")
-      .then((res) => res.json())
-      .then((data) => setHumans(data.data));
-  }, []);
   return (
     <Wrapper>
-      <ModalBox active={humanModal}>
-        <Modal className={"shell"}>
-          <h3 onClick={() => setHumanModal(false)}>X</h3>
-          <h2>You saw a trashy human?!?</h2>
-          <p>Tell us more...</p>
-          <p>Trashing level: </p>
-          <input
-            type="text"
-            maxLength="2"
-            placeholder="(0-99)"
-            value={trashingLevel}
-            onChange={handleTrash}
-          />
-          <p>Excitement level: </p>
-          <input
-            type="text"
-            maxLength="2"
-            placeholder="(0-99)"
-            value={excitementLevel}
-            onChange={handleExcitement}
-          />
-          <Button onClick={handleAddHuman}>Submit</Button>
-        </Modal>
-      </ModalBox>
+      <MapModal
+        trashingLevel={trashingLevel}
+        setTrashingLevel={setTrashingLevel}
+        excitementLevel={excitementLevel}
+        setExcitementLevel={setExcitementLevel}
+        humanModal={humanModal}
+        setHumanModal={setHumanModal}
+        handleAddHuman={handleAddHuman}
+      />
       <LoadScript
         googleMapsApiKey={process.env.REACT_APP_MAPS_KEY}
         onLoad={() => {
@@ -175,68 +155,10 @@ const Map = () => {
   );
 };
 
-const fadeIn = keyframes`
-  from {
-    transform: translateY(20px);
-    opacity: 0;
-  }
-  to {
-    transform: translateY(0);
-    opacity: 1;
-  }
-`;
 const P = styled.p`
   margin: 10px 25px;
   font-size: large;
 `;
-const ModalBox = styled.div`
-  position: absolute;
-  top: 0;
-  z-index: 2;
-  width: 100vw;
-  height: 100vh;
-  background-color: rgba(0, 0, 0, 0.4);
-  ${({ active }) =>
-    active === true
-      ? `
-    display: flex;
-  `
-      : `
-    display: none;
-  `}
-  align-items: center;
-  justify-content: center;
-  .shell {
-    animation: ${fadeIn} 500ms;
-  }
-`;
-const Modal = styled.div`
-  background-color: white;
-  border-radius: 8px;
-  padding: 10px;
-  position: relative;
-  padding: 18px;
-  h2 {
-    font-weight: bold;
-    font-size: large;
-    margin-bottom: 4px;
-  }
-  h3 {
-    position: absolute;
-    top: 5px;
-    right: 5px;
-    cursor: pointer;
-    font-weight: bold;
-  }
-  p {
-    margin-bottom: 4px;
-  }
-  input {
-    width: 40px;
-    margin: 10px 80px 10px 0;
-  }
-`;
-
 const Wrapper = styled.div`
   height: 500px;
   display: flex;
